Link Back/Front Registration (Admin + User) - update display Register form - Add button Logout Navbar + CurrentUser States
