+++ src/Components/footer/footer.jsx
@@ -1,14 +1,13 @@
-import './footer.css'
-import { Link } from 'react-router-dom'
-
+import "./footer.css";
+import { Link } from "react-router-dom";
 
 function Footer() {
   return (
     <div>
-<footer className="dark:bg-gray-900">
-    <div className="mx-auto w-full max-w-screen-xl p-4 py-6 lg:py-8">
-        <div className="md:flex md:justify-between">
-          <div className="mb-6 md:mb-0">
+      <footer className="dark:bg-gray-900">
+        <div className="mx-auto w-full max-w-screen-xl p-4 py-6 lg:py-8">
+          <div className="md:flex md:justify-between">
+            <div className="mb-6 md:mb-0">
               <a href="https://flowbite.com/" className="flex items-center">
                 <img
                   src="https://flowbite.com/docs/images/logo.svg"
@@ -19,22 +18,44 @@ function Footer() {
                   Flowbite
                 </span>
               </a>
+            </div>
+            <div>
+              <Link to="/Privacy" className="hover:underline m-2">
+                Privacy Policy
+              </Link>
+              <Link to="/Terms" className="hover:underline m-2">
+                Terms &amp; Conditions
+              </Link>
+            </div>
           </div>
-          <div>
-                          <Link to="/Privacy" className="hover:underline m-2">Privacy Policy</Link>
-                          <Link to="/Terms" className="hover:underline m-2">Terms &amp; Conditions</Link>
-          </div>
-      </div>
-      <hr className="my-6 border-gray-200 sm:mx-auto dark:border-gray-700 lg:my-8" />
-      <div className="sm:flex sm:items-center sm:justify-between">
-          <span className="text-sm text-primary-light sm:text-center dark:text-gray-400">© 2023 <a href="https://flowbite.com/" className="hover:underline">Flowbite™</a>. All Rights Reserved.
-          </span>
-          <div className="flex mt-4 sm:justify-center sm:mt-0">
-              <a href="#" className="text-gray-500 hover:text-gray-900 dark:hover:text-white">
-                  <svg className="w-4 h-4" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 8 19" fill='#BEE8FF'>
-                        <path fillRule="evenodd" d="M6.135 3H8V0H6.135a4.147 4.147 0 0 0-4.142 4.142V6H0v3h2v9.938h3V9h2.021l.592-3H5V3.591A.6.6 0 0 1 5.592 3h.543Z" clipRule="evenodd"/>
-                    </svg>
-                  <span className="sr-only">Facebook page</span>
+          <hr className="my-6 border-gray-200 sm:mx-auto dark:border-gray-700 lg:my-8" />
+          <div className="sm:flex sm:items-center sm:justify-between">
+            <span className="text-sm text-primary-light sm:text-center dark:text-gray-400">
+              © 2023{" "}
+              <a href="https://flowbite.com/" className="hover:underline">
+                Flowbite™
+              </a>
+              . All Rights Reserved.
+            </span>
+            <div className="flex mt-4 sm:justify-center sm:mt-0">
+              <a
+                href="#"
+                className="text-gray-500 hover:text-gray-900 dark:hover:text-white"
+              >
+                <svg
+                  className="w-4 h-4"
+                  aria-hidden="true"
+                  xmlns="http://www.w3.org/2000/svg"
+                  viewBox="0 0 8 19"
+                  fill="#BEE8FF"
+                >
+                  <path
+                    fillRule="evenodd"
+                    d="M6.135 3H8V0H6.135a4.147 4.147 0 0 0-4.142 4.142V6H0v3h2v9.938h3V9h2.021l.592-3H5V3.591A.6.6 0 0 1 5.592 3h.543Z"
+                    clipRule="evenodd"
+                  />
+                </svg>
+                <span className="sr-only">Facebook page</span>
               </a>
               <a
                 href="#"
