+++ src/Components/navbar/Nav.jsx
@@ -9,9 +9,19 @@ import {
 import { axiosPrivate } from "../../api/axios";
 import { useAtom } from "jotai";
 import { bearerToken } from "../../atom/atoms";
+import { useEffect, useState } from "react";
 
 const Nav = () => {
   const [token, setToken] = useAtom(bearerToken);
+  const [currentUser, setCurrentUser] = useState(false);
+
+  useEffect(() => {
+    if (token) {
+      setCurrentUser(true);
+    } else {
+      setCurrentUser(false);
+    }
+  }, [token])
 
   const logout = () => {
     console.log(token)
@@ -24,6 +34,7 @@ const Nav = () => {
       .then((response) => {
         console.log(response)
         setToken(null);
+        setCurrentUser(false);
       })
       .catch(error => {
         console.error('Logout failed:', error);
@@ -42,12 +53,18 @@ const Nav = () => {
         </NavbarItem>
       </NavbarContent>
       <NavbarContent justify="end">
-        <NavbarItem className="hidden lg:flex">
-          <Link to="/login">Se connecter</Link>
-        </NavbarItem>
         <NavbarItem>
+          <Link to="/register">S'inscrire</Link>
+        </NavbarItem>
+        {currentUser ? (
+          <NavbarItem>
           <Button onClick={logout}>Se déconnecter</Button>
         </NavbarItem>
+        ) : (
+          <NavbarItem className="hidden lg:flex">
+          <Link to="/login">Se connecter</Link>
+        </NavbarItem>
+        )}       
       </NavbarContent>
     </Navbar>
   );
