+++ src/Pages/Home/Home.jsx
@@ -1,7 +1,10 @@
+
+
 function Home() {
   return (
     <div>
       <h1>Home</h1>
+     
     </div>
   )
 }
