+++ src/Pages/Register/Register.css
